Split out action_diff code
