+++ tests/action_diffs.rs
@@ -1,8 +1,6 @@
 use bevy::{input::InputPlugin, prelude::*};
-use leafwing_input_manager::action_state::ActionDiffEvent;
-use leafwing_input_manager::{
-    action_state::ActionDiff, axislike::DualAxisData, prelude::*, systems::generate_action_diffs,
-};
+use leafwing_input_manager::action_diff::{ActionDiff, ActionDiffEvent};
+use leafwing_input_manager::{axislike::DualAxisData, prelude::*, systems::generate_action_diffs};
 
 #[derive(Actionlike, Clone, Copy, Debug, Reflect, PartialEq, Eq, Hash)]
 enum Action {
